chore: move files based on their functionality, plus cleanup
